% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdim.R
\name{convert_tool_data_into_matrix}
\alias{convert_tool_data_into_matrix}
\title{Convert tool prediction dataframe into a matrix with rows as miRNAs, columns as genes and scores as values}
\usage{
convert_tool_data_into_matrix(pred_data, miRNA_index, gene_index, score_index)
}
\arguments{
\item{pred_data}{The predictions in the form of a data frame}

\item{miRNA_index}{The column index of miRNA in the dataframe}

\item{gene_index}{The column index of gene in the dataframe}

\item{score_index}{The column index of scores in your dataframe}
}
\description{
Convert tool prediction dataframe into a matrix with rows as miRNAs, columns as genes and scores as values
}
