% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_scor.R
\name{ensemble_scores}
\alias{ensemble_scores}
\title{Functionality to calculate the ensemble scores for all the tools}
\usage{
ensemble_scores(tool_list, tool_names, performance_list)
}
\arguments{
\item{tool_list}{A list containing the data frames containing scores for all the tools}

\item{tool_names}{A list containing the names of all the tools in the dataframes}

\item{performance_list}{A list containing performance scores of all the tools}
}
\description{
Functionality to calculate the ensemble scores for all the tools
}
