% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_prpc.R
\name{expression_preprocessing}
\alias{expression_preprocessing}
\title{preprocessing expression data}
\usage{
expression_preprocessing(
  expr_data,
  missing_value_filter = 0,
  missing_value,
  variance_filter = 0,
  mean_filter = 0
)
}
\arguments{
\item{expr_data}{The input expression matrix containing expression values}

\item{missing_value_filter}{Genes/miRNAs with more than missing_value\% will be removed.}

\item{missing_value}{The expression value which represents the missing value in your expression matrix.}

\item{variance_filter}{Genes/miRNAs with variance less than variance_filter quantile will be removed.Default is 0.}

\item{mean_filter}{Genes/miRNAs with mean less than mean_filter quantile will be removed.Default is 0.}
}
\description{
Filter the expression data
}
