% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expr}
\alias{gene_expr}
\title{Gene Expression data}
\format{
A matrix containing a subset of gene expression data
}
\usage{
gene_expr
}
\description{
Gene Expression data
}
\keyword{datasets}
