% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_row_col.R
\name{gene_miRNA_id_conversion}
\alias{gene_miRNA_id_conversion}
\title{Convert gene names and miRNA names into ensembl ids and accession ids respectively}
\usage{
gene_miRNA_id_conversion(
  mat_data,
  gene_bool = TRUE,
  cgid,
  dgid = "ensembl_gene_id",
  mir_bool = TRUE,
  cmirid = " ",
  dmirid = "Accession"
)
}
\arguments{
\item{mat_data}{expression matrix or tool prediction matrix with columns as genes and rows as miRNAs}

\item{gene_bool}{A boolean value. If it is TRUE, gene ids will be converted.TRUE by default}

\item{cgid}{A string containing current gene id format.}

\item{dgid}{A string containing desired gene id format.'ensembl_gene_id' by default.}

\item{mir_bool}{A boolean value. If it is TRUE, miRNA ids will be converted.TRUE by default}

\item{cmirid}{A string containing current miRNA id format.empty by default.version will be automatically detected}

\item{dmirid}{A string containing desired miRNA id format.}
}
\description{
Convert gene names and miRNA names into ensembl ids and accession ids respectively
}
