% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mir_expr}
\alias{mir_expr}
\title{miRNA Expression data}
\format{
A matrix containing a subset of miRNA expression data
}
\usage{
mir_expr
}
\description{
miRNA Expression data
}
\keyword{datasets}
