% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mirtarbase}
\alias{mirtarbase}
\title{mirtarbase tool scores}
\format{
A matrix containing a subset of mirtarbase scores
}
\usage{
mirtarbase
}
\description{
mirtarbase tool scores
}
\keyword{datasets}
