% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mirtarbase7}
\alias{mirtarbase7}
\title{mirtarbase7 tool scores}
\format{
A matrix containing a subset of mirtarbase7 scores
}
\usage{
mirtarbase7
}
\description{
mirtarbase7 tool scores
}
\keyword{datasets}
