% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mirwalk_pred}
\alias{mirwalk_pred}
\title{mirwalk tool predictions}
\format{
A matrix containing a subset of mirwalk predictions
}
\usage{
mirwalk_pred
}
\description{
mirwalk tool predictions
}
\keyword{datasets}
