% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_model.R
\name{regression_results}
\alias{regression_results}
\title{Regression models on the expression data}
\usage{
regression_results(dependent_expr, independent_expr, model_choice, rsq_filter)
}
\arguments{
\item{dependent_expr}{input  preprocessed dependent expression matrix}

\item{independent_expr}{input preprocessed independent expression matrix}

\item{model_choice}{String input that specifies the regression model to be used.
'e','l','r' and 's' for elastic net, lasso, ridge and simple linear regression respectievly.}

\item{rsq_filter}{A filter to select models/genes using the r squared value.All genes with rsq < rsq_filter will be removed.}
}
\description{
Regression models on the expression data
}
