% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_coefficients.R
\name{regression_vs_tool}
\alias{regression_vs_tool}
\title{Benchmark tools by comparing regression against tools}
\usage{
regression_vs_tool(reg_mat, tool_mat_list, val_mat, threshold_valmir = 100)
}
\arguments{
\item{reg_mat}{A matrix containing the regression coefficients in the form of a matrix/dataframe.}

\item{val_mat}{A validation dataset containing the number of binding sites to validate the miRNAs.}

\item{threshold_valmir}{A threshold value to filter the miRNAs.miRNAs with less than threshold gene targets will be removed.
The default is 100.}

\item{vector_tools}{A list containing dataframes from different tools.}
}
\description{
Benchmark tools by comparing regression against tools
}
