% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_mirwalk}
\alias{sample_mirwalk}
\title{mirwalk tool scores}
\format{
A matrix containing a subset of mirwalk scores
}
\usage{
sample_mirwalk
}
\description{
mirwalk tool scores
}
\keyword{datasets}
