% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_simple_regression_results}
\alias{sample_simple_regression_results}
\title{correlation matrix}
\format{
A matrix containing the gene-miRNA expression correlation
}
\usage{
sample_simple_regression_results
}
\description{
correlation matrix
}
\keyword{datasets}
