% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_targetscan}
\alias{sample_targetscan}
\title{targetscan sample tool}
\format{
A matrix containing a subset of targetscan scores
}
\usage{
sample_targetscan
}
\description{
targetscan sample tool
}
\keyword{datasets}
