% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetscan_pred}
\alias{targetscan_pred}
\title{targetscan tool predictions}
\format{
A matrix containing a subset of targetscan predictions
}
\usage{
targetscan_pred
}
\description{
targetscan tool predictions
}
\keyword{datasets}
