% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_score.R
\name{tool_vs_regression}
\alias{tool_vs_regression}
\title{Benchmark various tools against regression using the scores.}
\usage{
tool_vs_regression(reg_mat, tool_mat_list, threshold = 0)
}
\arguments{
\item{reg_mat}{A matrix containing the regression coefficients in the form of a matrix/dataframe.}

\item{threshold}{A threshold coefficient value for coefficients.All the coefficients less than the threshold
will be changed to 1 and others will be 0. threshold = 0 by default.}

\item{vector_tools}{A list containing dataframes from different tools.}
}
\description{
Benchmark various tools against regression using the scores.
}
