% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_tool_vs_reg.R
\name{tool_vs_regression_using_filtered_miRNAs}
\alias{tool_vs_regression_using_filtered_miRNAs}
\title{Benchmark tools using filtered miRNAs which have a lot of experimentally validated gene targets by mirtarbase7}
\usage{
tool_vs_regression_using_filtered_miRNAs(
  reg_mat,
  tool_mat_list,
  val_mat,
  threshold_valmir = 100,
  threshold_coeff = 0
)
}
\arguments{
\item{reg_mat}{A matrix containing the regression coefficients in the form of a matrix/dataframe.}

\item{val_mat}{A validation dataset containing the number of binding sites to validate the miRNAs.}

\item{threshold_valmir}{A threshold value to filter the miRNAs.miRNAs with less than threshold gene targets will be removed.
The default is 100.}

\item{threshold_coeff}{A threshold coefficient value for coefficients.All the coefficients less than the threshold
will be changed to 1 and others will be 0. threshold = 0 by default.}

\item{vector_tools}{A list containing dataframes from different tools.}
}
\description{
Benchmark tools using filtered miRNAs which have a lot of experimentally validated gene targets by mirtarbase7
}
