% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_val_mir.R
\name{validated_miRNA_filter}
\alias{validated_miRNA_filter}
\title{Filter tooland regression matrices}
\usage{
validated_miRNA_filter(
  matrix_list,
  matrix_names,
  val_mat,
  threshold_valmir = 100
)
}
\arguments{
\item{matrix_list}{A list containing all the tool data and regression matrix}

\item{matrix_names}{A list of names of all the data sets in the list}

\item{val_mat}{A matrix containg experimentally validated gene-miRNA interactions.}

\item{threshold_valmir}{A threshold for filtering miRNAs. Only the miRNAs which have more than threshold_valmir
gene targets will be kept. It is 100 by default.}
}
\description{
Filter the tool and regression matrixes using experimentally validated data
}
