% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_common_interactions.R
\name{common_data}
\alias{common_data}
\title{extract common genes and miRNAs from all the tool prediction matrices and regression matrix}
\usage{
common_data(reg_mat, tools, filter = NA)
}
\arguments{
\item{reg_mat}{A matrix containg the regression coefficients}

\item{tools}{A list containing prediction matrices/dataframes from different tools.Each element in the list is a matrix/dataframe}

\item{filter}{Genes/miRNAs with more than 95\% values equal to filter will be removed for all the tools. NA by default.}
}
\description{
extract common genes and miRNAs from all the tool prediction matrices and regression matrix
}
