% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdim.R
\name{convert_tool_data_into_matrix}
\alias{convert_tool_data_into_matrix}
\title{dataframe to matrix}
\usage{
convert_tool_data_into_matrix(
  pred_data,
  miRNA_colname,
  gene_colname,
  score_colname
)
}
\arguments{
\item{pred_data}{The predictions in the form of a data frame}

\item{miRNA_colname}{The column name of miRNA in the dataframe}

\item{gene_colname}{The column name of gene in the dataframe}

\item{score_colname}{The column name of scores in your dataframe}
}
\description{
Convert tool prediction dataframe into a matrix with rows as miRNAs, columns as genes and scores as values
}
