% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_coefficients.R
\name{regression_vs_tool}
\alias{regression_vs_tool}
\title{Compare with Regresssion}
\usage{
regression_vs_tool(reg_mat, tool_mat_list, tool_names)
}
\arguments{
\item{reg_mat}{A matrix containing the regression coefficients in the form of a matrix/dataframe.}

\item{tool_mat_list}{A list containing dataframes from different tools.}

\item{tool_names}{A list of the names of tools in the same order as tool_mat_list}
}
\description{
Benchmark tools by comparing regression against tools
}
